backend/requirements.txt
Flask==3.0.0
SQLAlchemy==2.0.23
PyMySQL==1.1.0
gTTS==2.5.0
Pillow==10.1.0
pytest==7.4.3
hypothesis==6.92.1
Flask-CORS==4.0.0
python-dotenv==1.0.0
mutagen==1.47.0
alembic==1.13.1
